requirements.txt
pytorch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
numpy<2.0
tqdm
pandas
matplotlib==3.7.1
pytorch3d==0.7.8